Add Hydra.Snapshot module
