+++ hydra-node/hydra-node.cabal
@@ -105,6 +105,7 @@ library
     Hydra.Node
     Hydra.Node.Version
     Hydra.Options
+    Hydra.Snapshot
   other-modules:
     Paths_hydra_node
   build-depends: base
