+++ hydra-node/src/Hydra/HeadLogic.hs
@@ -28,6 +28,7 @@ import Hydra.Ledger (
   verify,
  )
 import Hydra.ClientInput (ClientInput (..))
+import Hydra.Snapshot (Snapshot (..), SnapshotNumber)
 
 data Event tx
   = ClientEvent (ClientInput tx)
@@ -50,47 +51,6 @@ deriving instance Tx tx => Show (Effect tx)
 deriving instance Tx tx => ToJSON (Effect tx)
 deriving instance Tx tx => FromJSON (Effect tx)
 
-type SnapshotNumber = Natural
-
-data Snapshot tx = Snapshot
-  { number :: SnapshotNumber
-  , utxo :: UTxO tx
-  , -- | The set of transactions that lead to 'utxo'
-    confirmed :: [tx]
-  }
-  deriving (Generic)
-
-deriving instance Tx tx => Eq (Snapshot tx)
-deriving instance Tx tx => Show (Snapshot tx)
-deriving instance Tx tx => Read (Snapshot tx)
-
-instance (Arbitrary tx, Arbitrary (UTxO tx)) => Arbitrary (Snapshot tx) where
-  arbitrary = genericArbitrary
-
-  -- NOTE: See note on 'Arbitrary (ClientInput tx)'
-  shrink s =
-    [ Snapshot (number s) utxo' confirmed'
-    | utxo' <- shrink (utxo s)
-    , confirmed' <- shrink (confirmed s)
-    ]
-
-instance Tx tx => SignableRepresentation (Snapshot tx) where
-  getSignableRepresentation = encodeUtf8 . show @Text
-
-instance Tx tx => ToJSON (Snapshot tx) where
-  toJSON s =
-    object
-      [ "snapshotNumber" .= number s
-      , "utxo" .= utxo s
-      , "confirmedTransactions" .= confirmed s
-      ]
-
-instance Tx tx => FromJSON (Snapshot tx) where
-  parseJSON = withObject "Snapshot" $ \obj ->
-    Snapshot
-      <$> (obj .: "snapshotNumber")
-      <*> (obj .: "utxo")
-      <*> (obj .: "confirmedTransactions")
 
 data ServerOutput tx
   = PeerConnected Party
@@ -231,9 +191,6 @@ instance (ToCBOR tx, ToCBOR (UTxO tx)) => ToCBOR (HydraMessage tx) where
     Connected host -> toCBOR ("Connected" :: Text) <> toCBOR host
     Disconnected host -> toCBOR ("Disconnected" :: Text) <> toCBOR host
 
-instance (ToCBOR tx, ToCBOR (UTxO tx)) => ToCBOR (Snapshot tx) where
-  toCBOR Snapshot{number, utxo, confirmed} = toCBOR number <> toCBOR utxo <> toCBOR confirmed
-
 instance (FromCBOR tx, FromCBOR (UTxO tx)) => FromCBOR (HydraMessage tx) where
   fromCBOR =
     fromCBOR >>= \case
@@ -243,9 +200,6 @@ instance (FromCBOR tx, FromCBOR (UTxO tx)) => FromCBOR (HydraMessage tx) where
       "Connected" -> Connected <$> fromCBOR
       "Disconnected" -> Disconnected <$> fromCBOR
       msg -> fail $ show msg <> " is not a proper CBOR-encoded HydraMessage"
-
-instance (FromCBOR tx, FromCBOR (UTxO tx)) => FromCBOR (Snapshot tx) where
-  fromCBOR = Snapshot <$> fromCBOR <*> fromCBOR <*> fromCBOR
 
 getParty :: HydraMessage msg -> Party
 getParty =
